--- onchain-academy/programs/solana-lms/apps/lms/app/(dashboard)/certificates/page.tsx
@@ -1,5 +1,0 @@
-import DashboardHome from "@/components/home";
-
-export default function Page() {
-  return <DashboardHome />;
-}
